Fixed physical attack property calc results
Since we didn't floor them, the client would use the raw values and sometimes list long floating point numbers in the stat window.

+++ system/scripts/zone/core/calc_character.cs
@@ -800,7 +800,7 @@ public class CharacterCalculationsScript : GeneralScript
 		value += byBuffs + byRateBuffs;
 
 		var max = SCR_Get_Character_MAXPATK(character);
-		return (float)Math2.Clamp(1, max, value);
+		return (int)Math2.Clamp(1, max, value);
 	}
 
 	/// <summary>
@@ -843,7 +843,7 @@ public class CharacterCalculationsScript : GeneralScript
 
 		value += byBuffs + byRateBuffs;
 
-		return (float)Math.Max(1, value);
+		return (int)Math.Max(1, value);
 	}
 
 	/// <summary>
